+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## 2020-07-15
+
+### Added
+
+- Integration tests for api auth backend
+
+## 2020-07-13
+
+### Added
+
+- Protect airflow api with custom hawk auth backend
+
 ## 2020-07-10
 
 ### Added